Refactor telefonoa into multiple modules
