+++ phonebook.yaml
@@ -1,5 +1,5 @@
-# Add up to 8 numbers corresponding for number 1-8 when hooked
+# Add up to 8 numbers corresponding for number 1-8 when off the hooked
 - name: Number 1
-  number: 123456789
+  number: 0419239384
 - name: Number 2
   number: 987654321
